Fixed Xcode "Value Conversion" warnings
Added explicit casts to existing implicit casts with potential data loss.

+++ 3rdparty/quazip/quazip/quaadler32.cpp
@@ -9,17 +9,17 @@ QuaAdler32::QuaAdler32()
 
 quint32 QuaAdler32::calculate(const QByteArray &data)
 {
-	return adler32( adler32(0L, Z_NULL, 0), (const Bytef*)data.data(), data.size() );
+	return (uint)adler32( adler32(0L, Z_NULL, 0), (const Bytef*)data.data(), data.size() );
 }
 
 void QuaAdler32::reset()
 {
-	checksum = adler32(0L, Z_NULL, 0);
+	checksum = (uint)adler32(0L, Z_NULL, 0);
 }
 
 void QuaAdler32::update(const QByteArray &buf)
 {
-	checksum = adler32( checksum, (const Bytef*)buf.data(), buf.size() );
+	checksum = (uint)adler32( checksum, (const Bytef*)buf.data(), buf.size() );
 }
 
 quint32 QuaAdler32::value()
