+++ sleepyhead/SleepLib/loader_plugins/cms50_loader.cpp
@@ -607,7 +607,7 @@ bool CMS50Loader::readSpoRFile(QString path)
 
         in >> samples;
 
-        int remainder = filesize - pos;
+        int remainder = (int)(filesize - pos);
 
         bytes_per_record = remainder / samples;
         qDebug() << samples << "samples of" << bytes_per_record << "bytes each";
